float fix

+++ main.go
@@ -23,7 +23,7 @@ func main() {
 	total_clients := []uint32{20}
 	shufflers := []uint32{3}
 	// total_clients := []uint32{100, 80, 60, 40, 20}
-	sitout_percent := 0.1
+	sitout_percent := float64(0.1)
 	for _, total_client := range total_clients {
 		for _, shuffler := range shufflers {
 			run_task := datastruct.RunTask{
